Backslashes should've valid uses too, even in safe-mode.
Must trust to the user with setting the paths, I suppose. We can't expect that a backslash shouldn't be there and risking the backup process writing over a wrong file isn't an option.

This whole PHP safe-mode restriction (vs. shell) is bit hard to tackle and handle in somewhat elegant manner.